Avoid a 64-to-32-bit conversion warning in JavaHL.
[in subversion/bindings/javahl]
* native/SVNClient.h, native/SVNClient.cpp (SVNClient::logMessages):
   Change the type of the 'limit' parameter from long to int,
   to match the prototype of svn_client_log5.
* native/org_apache_subversion_javahl_SVNClient.cpp
  (Java_org_apache_subversion_javahl_SVNClient_logMesages):
   Verify that the jlimit argument is within the range of the C int
   type, and cast it accordingly in the call to SVNClient::logMessages.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1603363 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/bindings/javahl/native/SVNClient.h
@@ -159,7 +159,7 @@ class SVNClient :public SVNBase
                    std::vector<RevisionRange> &ranges, bool stopOnCopy,
                    bool discoverPaths, bool includeMergedRevisions,
                    StringArray &revProps,
-                   long limit, LogMessageCallback *callback);
+                   int limit, LogMessageCallback *callback);
   jobject getVersionExtended(bool verbose);
   jstring getAdminDirectoryName();
   jboolean isAdminDirectory(const char *name);
